- Refactor files - Add Gif Logo
