--- src/components/Tempelate/index.tsx
@@ -1,7 +1,0 @@
-const Tempelate = () => {
-  return (
-    <div />
-  );
-}
-
-export default Tempelate;
