+++ src/pages/Landing/SectionFour/index.scss
@@ -1,0 +1,37 @@
+@import 'src/assets/styles/variables';
+@import 'src/assets/styles/functions';
+
+.section {
+  &__four {
+    background:      url('../../../../public/images/section-four-back.png') no-repeat top center;
+    background-size: cover;
+    padding:         860px 0 40px;
+    text-align:      center;
+
+    .coming-soon-text {
+      color:          $meta-white;
+      font-family:    Montserrat;
+      font-size:      rem-calc(34);
+      font-weight:    300;
+      letter-spacing: 0.58em;
+      margin-bottom:  26px;
+      text-transform: uppercase;
+    }
+  }
+}
+
+@media only screen and (max-width: 944px) {
+  .section {
+    &__four {
+      padding: 590px 0 40px;
+
+      .coming-soon-text {
+        font-size: rem-calc(18);
+      }
+
+      .twenty-twenty-two {
+        width: 160px;
+      }
+    }
+  }
+}
