+++ src/pages/Landing/SectionOne/index.scss
@@ -1,0 +1,85 @@
+@import 'src/assets/styles/variables';
+@import 'src/assets/styles/functions';
+
+.section {
+  &__one {
+    align-items:      center;
+    background:       url('../../../../public/images/section-one-back.png') no-repeat top center;
+    background-color: $meta-black;
+    background-size:  cover;
+    display:          flex;
+    flex-direction:   column;
+    justify-content:  center;
+    padding:          299px 0;
+
+    .gif-wrapper {
+      align-items:     center;
+      background:      url('../../../assets/images/gif-ring.png') no-repeat center center;
+      background-size: contain;
+      display:         flex;
+      justify-content: center;
+      margin-bottom:   70px;
+      position:        relative;
+      width:           786px;
+
+      .characters {
+        position: absolute;
+        top:      -85px;
+        width:    303px;
+      }
+
+      .image {
+        height: auto;
+        width:  100%;
+      }
+    }
+
+    .tag-line {
+      color:       $meta-white;
+      font-family: Gilroy;
+      font-size:   rem-calc(20);
+      font-weight: 800;
+      max-width:   971px;
+      padding:     0 29px;
+      text-align:  center;
+    }
+  }
+}
+
+@media only screen and (max-width: 1440px) {
+  .section {
+    &__one {
+      padding-bottom: 100px;
+
+      .gif-wrapper {
+        width: 576px;
+
+        .characters {
+          width: 243px;
+        }
+      }
+    }
+  }
+}
+
+@media only screen and (max-width: 944px) {
+  .section {
+    &__one {
+      padding: 210px 29px 100px;
+
+      .tag-line {
+        font-size: rem-calc(16);
+        padding:   0;
+      }
+
+      .gif-wrapper {
+        width: 300px;
+
+        .characters {
+          top:   -40px;
+          width: 123px;
+        }
+      }
+    }
+  }
+}
