+++ src/pages/Landing/SectionThree/index.scss
@@ -1,0 +1,211 @@
+@import 'src/assets/styles/variables';
+@import 'src/assets/styles/functions';
+
+.section {
+  &__three {
+    align-items:     center;
+    color:           $meta-white;
+    display:         flex;
+    flex-direction:  column;
+    justify-content: center;
+    text-align:      center;
+
+    .head {
+      font-family: Montserrat;
+      font-size:   rem-calc(64);
+      font-weight: 800;
+      margin:      50px 0 20px;
+    }
+
+    .description {
+      font-family:   Gilroy;
+      font-size:     rem-calc(20);
+      margin-bottom: 20px;
+      max-width:     460px;
+    }
+
+    .group-menu {
+      font-family: Gilroy;
+      list-style:  none;
+      margin:      50px 0 150px;
+
+      &__item {
+        display:        inline-block;
+        font-size:      rem-calc(18);
+        font-weight:    bold;
+        margin:         0 10px;
+        padding:        3px 12px;
+        text-transform: uppercase;
+
+        &:hover {
+          color:  $meta-green;
+          cursor: pointer;
+        }
+
+        &--active {
+          border:        1px solid $meta-green;
+          border-radius: 3px;
+          color:         $meta-green;
+        }
+      }
+    }
+
+    .item-list {
+      display:         flex;
+      justify-content: space-evenly;
+      max-width:       1303px;
+      width:           100%;
+    }
+
+    .character-card {
+      align-items:     center;
+      background:      linear-gradient(0deg, rgba(0, 0, 0, 0.2), rgba(0, 0, 0, 0.2));
+      border:          11px solid $meta-green;
+      border-radius:   8px;
+      display:         flex;
+      height:          445px;
+      justify-content: center;
+      position:        relative;
+      width:           395px;
+
+      &:nth-child(2) {
+        margin: 0 59px;
+      }
+
+      .image-cont {
+        align-items:     center;
+        display:         flex;
+        justify-content: center;
+        position:        relative;
+
+        .character-image {
+          height:   266px;
+          position: absolute;
+          width:    auto;
+        }
+
+        .circle {
+          background:    $meta-green;
+          border-radius: 100%;
+          height:        229px;
+          width:         229px;
+        }
+      }
+
+      .name {
+        bottom:         22px;
+        font-family:    Montserrat;
+        font-size:      rem-calc(25);
+        font-weight:    bold;
+        position:       absolute;
+        text-transform: uppercase;
+      }
+    }
+  }
+}
+
+@media only screen and (max-width: 1304px) {
+  .section {
+    &__three {
+      .character-card {
+        width: 32%;
+
+        &:nth-child(2) {
+          margin: 0;
+        }
+
+        .image-cont {
+          .character-image {
+            height: auto;
+            width:  120%;
+
+          }
+        }
+      }
+    }
+  }
+}
+
+@media only screen and (max-width: 1440px) {
+  .section {
+    &__three {
+      .head {
+        font-size: rem-calc(32);
+      }
+
+      .description {
+        font-size: rem-calc(16);
+      }
+
+      .group-menu {
+        margin: 50px 0;
+
+        .group-menu__item {
+          font-size: rem-calc(16);
+        }
+      }
+
+      .character-card {
+        height: 355px;
+        width:  29%;
+
+        .image-cont {
+
+          .circle {
+            height: 189px;
+            width:  189px;
+          }
+        }
+      }
+    }
+  }
+}
+
+@media only screen and (max-width: 944px) {
+  .section {
+    &__three {
+
+      .head {
+        margin-top: 0;
+      }
+
+      .description {
+        padding: 0 29px;
+      }
+
+      .group-menu {
+        align-items:     center;
+        display:         flex;
+        justify-content: space-between;
+        margin:          30px 0 0;
+        padding:         0 29px;
+
+        &__item {
+          margin: 0;
+        }
+      }
+
+      .item-list {
+        align-items:    center;
+        flex-direction: column;
+        padding:        0 29px;
+      }
+
+      .character-card {
+        margin:    70px 0;
+        max-width: 395px;
+        width:     100%;
+
+        .image-cont {
+          .character-image {
+
+          }
+        }
+
+        &:last-of-type {
+          margin-bottom: 0;
+        }
+      }
+    }
+  }
+}
