+++ src/pages/Landing/SectionTwo/index.scss
@@ -1,0 +1,287 @@
+@import 'src/assets/styles/variables';
+@import 'src/assets/styles/functions';
+
+.section {
+  &__two {
+    .section-image {
+      height: 613px;
+    }
+
+    #section-two-back {
+      margin-top: 52px;
+
+      .bot-image {
+        margin-left: 116px;
+        width:       29%;
+      }
+    }
+
+    #section-three-back {
+      padding: 0 370px 0 128px;
+
+      .bot-image {
+        position: absolute;
+        right:    20%;
+        top:      -107px;
+        width:    13.2%;
+      }
+    }
+
+    #section-four-back {
+      padding: 317px 0 0 128px;
+
+      .bot-image {
+        width: 44%;
+      }
+
+      .text-content {
+        margin-top: 100px;
+      }
+    }
+
+    #section-five-back {
+      justify-content: space-between;
+      padding-left:    128px;
+
+      .bot-image {
+        margin: 56px 60px 0 0;
+        width:  50%;
+      }
+    }
+
+    .bot-card {
+      color:       $meta-white;
+      display:     flex;
+      font-family: Montserrat;
+      margin:      60px 0 130px;
+      position:    relative;
+
+      .bot-image {
+        height: fit-content;
+      }
+
+      .text-content {
+        width: auto;
+      }
+
+      &:nth-child(2n) {
+        flex-direction:  row-reverse;
+        justify-content: flex-end;
+      }
+
+      &:nth-child(1), &:nth-child(3) {
+        .text-content {
+          margin-left:  80px;
+          margin-right: 20px;
+        }
+      }
+
+      .text-content {
+        .head {
+          font-size:     rem-calc(64);
+          font-weight:   800;
+          margin-bottom: 61px;
+          white-space:   break-spaces;
+        }
+
+        .description {
+          font-family: Gilroy;
+          font-size:   rem-calc(20);
+          max-width:   753px;
+          white-space: break-spaces;
+        }
+      }
+
+      &#section-two-back {
+        .text-content {
+          .description {
+            max-width: 728px;
+          }
+        }
+      }
+
+      &#section-three-back {
+        .text-content {
+          .description {
+            max-width: 806px;
+          }
+        }
+      }
+
+      &#section-five-back {
+        .text-content {
+          .description {
+            max-width: 559px;
+          }
+        }
+      }
+    }
+  }
+}
+
+@media only screen and (max-width: 1440px) {
+  .section {
+    &__two {
+      .bot-card {
+        justify-content: space-around !important;
+        margin:          90px 0 !important;
+        padding:         0 40px !important;
+
+        &#section-two-back {
+          .text-content {
+            .description {
+              max-width: 583px;
+            }
+          }
+        }
+
+        &#section-three-back {
+          .text-content {
+            .description {
+              max-width: 643px;
+            }
+          }
+        }
+
+        &#section-four-back {
+          .text-content {
+            .description {
+              max-width: 602px;
+            }
+          }
+        }
+
+        &#section-five-back {
+          .text-content {
+            .description {
+              max-width: 443px;
+            }
+          }
+        }
+
+        .bot-image {
+          margin:   0 !important;
+          position: initial !important;
+        }
+
+        .text-content {
+          margin: 0 !important;
+
+          .head {
+            font-size: rem-calc(32) !important;
+          }
+
+          .description {
+            font-size: rem-calc(16) !important;
+          }
+        }
+      }
+
+      #section-two-back {
+        .bot-image {
+          width: 30% !important;
+        }
+      }
+
+      #section-three-back {
+
+        .bot-image {
+          position: relative !important;
+          top:      -70px !important;
+        }
+      }
+    }
+  }
+}
+
+@media only screen and (max-width: 1328px) {
+  .section {
+    &__two {
+      #section-three-back {
+        .bot-image {
+          top: 0;
+        }
+      }
+    }
+  }
+}
+
+@media only screen and (max-width: 1577px) {
+  .section {
+    &__two {
+      #section-three-back {
+
+        .bot-image {
+          right: 80px;
+        }
+      }
+
+      #section-four-back {
+        padding: 77px 0 0 128px;
+      }
+    }
+  }
+}
+
+@media only screen and (max-width: 944px) {
+  .section {
+    &__two {
+      .bot-card {
+        align-items:    center;
+        flex-direction: column !important;
+        margin:         40px 0 90px !important;
+        padding:        0 29px !important;
+
+        &:first-of-type {
+          margin-top: 0 !important;
+        }
+
+        .bot-image {
+          height:        fit-content;
+          margin-bottom: 30px !important;
+          width:         100% !important;
+        }
+
+        &#section-two-back {
+          .bot-image {
+            width: 35% !important;
+          }
+        }
+
+        &#section-three-back {
+          .bot-image {
+            position: initial !important;
+            width:    17% !important;
+          }
+        }
+
+        &#section-four-back {
+          .bot-image {
+            width: 44% !important;
+          }
+        }
+
+        &#section-five-back {
+          .bot-image {
+            width: 58% !important;
+          }
+        }
+
+        .text-content {
+          margin:        0 auto;
+          margin-bottom: 15px !important;
+          text-align:    center;
+
+          .head {
+            margin-bottom: 21px;
+            white-space:   initial;
+          }
+
+          .description {
+            width: 100% !important;
+          }
+        }
+      }
+    }
+  }
+}
